Bugfix(es): 3 top nav adjacent bugs we want to cherrypick into the 1.46.0 release (#11663)
This PR fixes a few bugs:
1. [important] In order to get the mobile sidebar overlapping the app
functionality working in #11349 , I removed a `position: "relative"`.
This had the unintended consequence of making it so that the header
didn't use that div for calculating its position and as a side effect,
made it so that whenever the sidebar opened, it overlapped the top nav.
Bad. This PR fixes that, by conditionally the `position: "relative"` to
only apply in mobile mode.
2. [might as well include] There would've been another bug in fixing the
above bug because we were using the `sm` breakpoint to determine when to
put the top nav items into the sidebar, but `md` to determine when the
sidebar should overlap the app. This PR makes it so both of these things
use the `md` breakpoint sidestepping that bug.
3. [important] There was a bad merge with
https://github.com/streamlit/streamlit/pull/11412 and
https://github.com/streamlit/streamlit/pull/11349/ where the former
changed a method that the latter had moved to a different file so git
didn't pick up on the conflict and it went unaddressed. This PR
addresses it and coalesces them.

~~4. [might as well include] In the process of fixing the above, I
realized the sidebar being long enough to scroll would make it so that
you can't collapse the sidebar as the sidebar header would go off the
top of the screen. This PR also fixes that issue so that they always
stay visible.~~ <-- backing this change out as it caused a difficult to
debug issue in playwright tests

none

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- E2E Tests
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.